添加秒杀倒计时、秒杀中和秒杀结束相关业务，以及秒杀成功和失败的业务逻辑

+++ src/main/java/com/glw/miaosha/doman/Goods.java
@@ -1,6 +1,5 @@
 package com.glw.miaosha.doman;
 
-import java.math.BigDecimal;
 
 public class Goods {
     private Long id;
@@ -11,7 +10,7 @@ public class Goods {
 
     private String goodsImg;
 
-    private BigDecimal goodsPrice;
+    private Long goodsPrice;
 
     private Integer goodsStock;
 
@@ -49,11 +48,11 @@ public class Goods {
         this.goodsImg = goodsImg == null ? null : goodsImg.trim();
     }
 
-    public BigDecimal getGoodsPrice() {
+    public Long getGoodsPrice() {
         return goodsPrice;
     }
 
-    public void setGoodsPrice(BigDecimal goodsPrice) {
+    public void setGoodsPrice(Long goodsPrice) {
         this.goodsPrice = goodsPrice;
     }
